+++ examples/graphiql-2-rfc-context/src/api/hooks/useValueRef.ts
@@ -11,7 +11,7 @@ import { useEffect, useRef, MutableRefObject } from 'react';
 /**
  * useValueRef
  *
- * Returns a reference to a given value. Automatically updates the value of the refrence when the value updates
+ * Returns a reference to a given value. Automatically updates the value of the reference when the value updates
  *
  */
 export default function useValueRef<T>(value: T): MutableRefObject<T> {