+++ examples/graphiql-2-rfc-context/src/components/HeaderEditor.tsx
@@ -148,7 +148,7 @@ export class HeaderEditor extends React.Component<HeaderEditorProps> {
       return;
     }
 
-    // Ensure the changes caused by this update are not interpretted as
+    // Ensure the changes caused by this update are not interpreted as
     // user-input changes which could otherwise result in an infinite
     // event loop.
     this.ignoreChangeEvent = true;