+++ packages/graphql-language-service/src/parser/Rules.ts
@@ -37,7 +37,7 @@ export const LexRules = {
   // Combines the IntValue and FloatValue tokens.
   Number: /^-?(?:0|(?:[1-9][0-9]*))(?:\.[0-9]*)?(?:[eE][+-]?[0-9]+)?/,
 
-  // Note the closing quote is made optional as an IDE experience improvment.
+  // Note the closing quote is made optional as an IDE experience improvement.
   String: /^(?:"""(?:\\"""|[^"]|"[^"]|""[^"])*(?:""")?|"(?:[^"\\]|\\(?:"|\/|\\|b|f|n|r|t|u[0-9a-fA-F]{4}))*"?)/,
 
   // Comments consume entire lines.